requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
sentence-transformers
requests
pillow
python-dotenv
openai
